Create Challenge Project: Credit Card Checker (Using Luhn Algorithm)
Stage 1:
In this project, we create a function, validateCred() that has a parameter of an array. 
The purpose of validateCred() is to return true when an array contains digits of a valid credit card number and false when it is invalid. (This function should NOT mutate the values of the original array.)

To find out if a credit card number is valid or not, we use the Luhn algorithm. The Luhn algorithm is a series of mathematical calculations used to validate certain identification numbers, e.g. credit card numbers. The calculations in the Luhn algorithm can be broken down as the following steps:

1. Starting from the farthest digit to the right, AKA the check digit, iterate to the left.

2. As we iterate to the left, every other digit is doubled (the check digit is not doubled). 

3. If the number is greater than 9 after doubling, subtract 9 from its value.

4. Sum up all the digits in the credit card number.

5. If the sum modulo 10 is 0 (if the sum divided by 10 has a remainder of 0) then the number is valid, otherwise, it’s invalid.

Stage 2:
We then create another function, findInvalidCards() that has one parameter for a nested array of credit card numbers. The role of findInvalidCards() is to check through the nested array for which numbers are invalid, and return another nested array of invalid cards.

Stage 3:
After finding all the invalid credit card numbers, we want to identify the credit card companies that have possibly issued these faulty numbers.

We create a function, idInvalidCardCompanies() that has one parameter for a nested array of invalid numbers and returns an array of companies. 
The function should return an array of companies that have mailed out cards with invalid numbers. 
(This array should NOT contain duplicates, i.e. even if there are two invalid Visa cards, "Visa" should only appear once in the array.)